rename(lottotest, lottostest): test -> domainTest 패키지 변경
