+++ src/test/java/lotto/domainTest/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domainTest;
 
 import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
